requirements.txt
tensorflow-macos==2.16.2
numpy==1.26.4
librosa==0.10.1
scikit-learn==1.5.0
pandas==2.2.2
tqdm==4.66.1